refactor: move examples into own folder
