+++ examples/conv2d_mnist/params.rs
