+++ src/lib.rs
@@ -5,14 +5,13 @@ pub mod affine1d;
 //mod apiexperiment;
 //pub mod cnvrl;
 pub mod cnvrl_generic;
-mod conv2d_mnist_example;
 pub mod eltwise;
 pub mod fieldutils;
 pub mod inputlayer;
 //mod softmax;
 //mod layertrait;
 //mod mlprescale;
-mod moreparams;
+// mod moreparams;
 //mod mvmul2; // debug overflow in V1
 //mod relu8; //move tests
 pub mod tensorutils;
